ManiControl/requirements.txt
numpy
numpy-quaternion
pybullet
open3d
